+++ target_chains/ethereum/contracts/contracts/pulse/PulseUpgradeable.sol
@@ -25,7 +25,7 @@ contract PulseUpgradeable is
         address pythAddress,
         address defaultProvider,
         bool prefillRequestStorage,
-        uint256 exclusivityPeriodSeconds
+        uint32 exclusivityPeriodSeconds
     ) external initializer {
         require(owner != address(0), "owner is zero address");
         require(admin != address(0), "admin is zero address");
